Make charts re-usable
