+++ src/components/ChartModal/BarChartConfig/styles.css
