+++ src/components/ChartModal/HistogramConfig/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useMemo } from 'react';
+import React, { useCallback, useState } from 'react';
 import {
     randomString,
     isFalsyString,
@@ -12,70 +12,24 @@ import Button from '#components/Button';
 import TextInput from '#components/TextInput';
 import ColorInput from '#components/ColorInput';
 import NumberInput from '#components/NumberInput';
-import { ExtractKeys } from '#utils/common';
 import { tableauColors } from '#utils/constants';
-import { Indicator } from '#types';
 
-import { ExtendedFiveW } from '../../../../useExtendedFiveW';
-import { HistogramSettings } from '../../types';
+import { HistogramSettings, NumericOption } from '#types';
 import styles from './styles.css';
 
-interface NumericOption {
-    key: string;
-    title: string;
-    valueSelector: (value: ExtendedFiveW) => number;
-    dependency?: number;
-    category: string;
-}
-
-const numericOptions: NumericOption[] = [
-    {
-        key: 'allocatedBudget',
-        title: 'Allocated Budget',
-        valueSelector: item => item.allocatedBudget,
-        category: 'DFID Data',
-    },
-    {
-        key: 'programCount',
-        title: '# of programs',
-        valueSelector: item => item.programCount,
-        category: 'DFID Data',
-    },
-    {
-        key: 'componentCount',
-        title: '# of components',
-        valueSelector: item => item.componentCount,
-        category: 'DFID Data',
-    },
-    {
-        key: 'partnerCount',
-        title: '# of partners',
-        valueSelector: item => item.partnerCount,
-        category: 'DFID Data',
-    },
-    {
-        key: 'sectorCount',
-        title: '# of sectors',
-        valueSelector: item => item.sectorCount,
-        category: 'DFID Data',
-    },
-];
-
-const keySelector = (item: NumericOption) => item.key;
-const labelSelector = (item: NumericOption) => item.title;
-const groupSelector = (item: NumericOption) => item.category;
-
-interface Props {
-    onSave: (settings: HistogramSettings<ExtendedFiveW>) => void;
-    indicatorList: Indicator[] | undefined;
+interface Props<T> {
+    onSave: (settings: HistogramSettings<T>) => void;
     className?: string;
+    options: NumericOption<T>[];
+    // keySelector: (item: T) => string;
 }
 
-function HistogramConfig(props: Props) {
+function HistogramConfig<T>(props: Props<T>) {
     const {
         onSave,
-        indicatorList,
         className,
+        options,
+        // keySelector: primaryKeySelector,
     } = props;
 
     const [error, setError] = useState<string | undefined>(undefined);
@@ -85,27 +39,9 @@ function HistogramConfig(props: Props) {
     const [color, setColor] = useState(() => getRandomFromList(tableauColors));
     const [binCount, setBinCount] = useState('10');
 
-    const options: NumericOption[] = useMemo(
-        () => {
-            if (!indicatorList) {
-                return numericOptions;
-            }
-            return [
-                ...numericOptions,
-                ...indicatorList.map(indicator => ({
-                    key: `indicator_${indicator.id}`,
-                    title: indicator.fullTitle,
-                    // FIXME: we should have certain thing for this
-                    valueSelector: (item: ExtendedFiveW) => item.indicators[indicator.id] || 0,
-
-                    category: indicator.category,
-
-                    dependency: indicator.id,
-                })),
-            ];
-        },
-        [indicatorList],
-    );
+    const keySelector = (item: NumericOption<T>) => item.key;
+    const labelSelector = (item: NumericOption<T>) => item.title;
+    const groupSelector = (item: NumericOption<T>) => item.category;
 
     const handleSave = useCallback(
         () => {
@@ -147,7 +83,7 @@ function HistogramConfig(props: Props) {
                 ? [option.dependency]
                 : undefined;
 
-            const settings: HistogramSettings<ExtendedFiveW> = {
+            const settings: HistogramSettings<T> = {
                 id: chartId,
                 type: 'histogram',
                 title,
