+++ src/components/ChartModal/HistogramConfig/styles.css
