+++ src/components/PolyChart/Histogram/styles.css
