+++ src/components/PolyChart/PieChart/styles.css
